+++ test/index.js
@@ -2,13 +2,13 @@
 
 const { throws, strictEqual } = require('assert');
 const { spawnSync } = require('child_process');
-const path = require('path');
+const { join } = require('path');
 const snapshot = require('snapshot-assertion');
 const TestDirector = require('../lib/TestDirector.js');
 const simulatePublishedTraces = require('./simulatePublishedTraces.js');
 
-const FIXTURES_PATH = path.join(__dirname, 'fixtures');
-const SNAPSHOTS_PATH = path.join(__dirname, 'snapshots');
+const FIXTURES_PATH = join(__dirname, 'fixtures');
+const SNAPSHOTS_PATH = join(__dirname, 'snapshots');
 const NODE_VERSION_MAJOR = parseInt(process.versions.node.split('.')[0]);
 
 const tests = [
@@ -41,14 +41,14 @@ const tests = [
     console.info('Test: Test passes.');
 
     const { stdout, stderr, status, error } = spawnSync('node', [
-      path.join(FIXTURES_PATH, 'passes'),
+      join(FIXTURES_PATH, 'passes'),
     ]);
 
     if (error) throw error;
 
     await snapshot(
       stdout.toString(),
-      path.join(SNAPSHOTS_PATH, 'test-passes-stdout.txt')
+      join(SNAPSHOTS_PATH, 'test-passes-stdout.txt')
     );
     strictEqual(stderr.toString(), '');
     strictEqual(status, 0);
@@ -58,14 +58,14 @@ const tests = [
     console.info('Test: Test with console output.');
 
     const { stdout, stderr, status, error } = spawnSync('node', [
-      path.join(FIXTURES_PATH, 'output'),
+      join(FIXTURES_PATH, 'output'),
     ]);
 
     if (error) throw error;
 
     await snapshot(
       stdout.toString(),
-      path.join(SNAPSHOTS_PATH, 'test-outputs-stdout.txt')
+      join(SNAPSHOTS_PATH, 'test-outputs-stdout.txt')
     );
     strictEqual(stderr.toString(), '');
     strictEqual(status, 0);
@@ -75,14 +75,14 @@ const tests = [
     console.info('Test: Awaits tests in sequence.');
 
     const { stdout, stderr, status, error } = spawnSync('node', [
-      path.join(FIXTURES_PATH, 'awaits'),
+      join(FIXTURES_PATH, 'awaits'),
     ]);
 
     if (error) throw error;
 
     await snapshot(
       stdout.toString(),
-      path.join(SNAPSHOTS_PATH, 'test-sequence-stdout.txt')
+      join(SNAPSHOTS_PATH, 'test-sequence-stdout.txt')
     );
     strictEqual(stderr.toString(), '');
     strictEqual(status, 0);
@@ -92,22 +92,18 @@ const tests = [
     console.info('Test: Test fails.');
 
     const { stdout, stderr, status, error } = spawnSync('node', [
-      path.join(FIXTURES_PATH, 'fails'),
+      join(FIXTURES_PATH, 'fails'),
     ]);
 
     if (error) throw error;
 
     await snapshot(
       stdout.toString(),
-      path.join(SNAPSHOTS_PATH, 'test-fails-stdout.txt')
+      join(SNAPSHOTS_PATH, 'test-fails-stdout.txt')
     );
     await snapshot(
       simulatePublishedTraces(stderr.toString()),
-      path.resolve(
-        __dirname,
-        'snapshots',
-        `test-fails-stderr-node-v${NODE_VERSION_MAJOR}.txt`
-      )
+      join(SNAPSHOTS_PATH, `test-fails-stderr-node-v${NODE_VERSION_MAJOR}.txt`)
     );
     strictEqual(status, 1);
   },
@@ -116,22 +112,18 @@ const tests = [
     console.info('Test: Nested tests.');
 
     const { stdout, stderr, status, error } = spawnSync('node', [
-      path.join(FIXTURES_PATH, 'nested'),
+      join(FIXTURES_PATH, 'nested'),
     ]);
 
     if (error) throw error;
 
     await snapshot(
       stdout.toString(),
-      path.join(SNAPSHOTS_PATH, 'test-nested-stdout.txt')
+      join(SNAPSHOTS_PATH, 'test-nested-stdout.txt')
     );
     await snapshot(
       simulatePublishedTraces(stderr.toString()),
-      path.resolve(
-        __dirname,
-        'snapshots',
-        `test-nested-stderr-node-v${NODE_VERSION_MAJOR}.txt`
-      )
+      join(SNAPSHOTS_PATH, `test-nested-stderr-node-v${NODE_VERSION_MAJOR}.txt`)
     );
     strictEqual(status, 1);
   },